+++ data_utils/S3DISDataLoader.py
@@ -33,7 +33,7 @@ class S3DISDataset(Dataset):
             self.room_points.append(points), self.room_labels.append(labels)
             self.room_coord_min.append(coord_min), self.room_coord_max.append(coord_max)
             num_point_all.append(labels.size)
-        labelweights = labelweights.astype(np.float32)
+        labelweights = labelweights.astype(float)
         labelweights = labelweights / np.sum(labelweights)
         self.labelweights = np.power(np.amax(labelweights) / labelweights, 1 / 3.0)
         print(self.labelweights)
@@ -114,7 +114,7 @@ class ScannetDatasetWholeScene():
             tmp, _ = np.histogram(seg, range(14))
             self.scene_points_num.append(seg.shape[0])
             labelweights += tmp
-        labelweights = labelweights.astype(np.float32)
+        labelweights = labelweights.astype(float)
         labelweights = labelweights / np.sum(labelweights)
         self.labelweights = np.power(np.amax(labelweights) / labelweights, 1 / 3.0)
