+++ models/pointnet_utils.py
@@ -36,7 +36,7 @@ class STN3d(nn.Module):
         x = F.relu(self.bn5(self.fc2(x)))
         x = self.fc3(x)
 
-        iden = Variable(torch.from_numpy(np.array([1, 0, 0, 0, 1, 0, 0, 0, 1]).astype(np.float32))).view(1, 9).repeat(
+        iden = Variable(torch.from_numpy(np.array([1, 0, 0, 0, 1, 0, 0, 0, 1]).astype(float))).view(1, 9).repeat(
             batchsize, 1)
         if x.is_cuda:
             iden = iden.cuda()
@@ -76,7 +76,7 @@ class STNkd(nn.Module):
         x = F.relu(self.bn5(self.fc2(x)))
         x = self.fc3(x)
 
-        iden = Variable(torch.from_numpy(np.eye(self.k).flatten().astype(np.float32))).view(1, self.k * self.k).repeat(
+        iden = Variable(torch.from_numpy(np.eye(self.k).flatten().astype(np.float))).view(1, self.k * self.k).repeat(
             batchsize, 1)
         if x.is_cuda:
             iden = iden.cuda()
